Disable CEngine.UpdateDPI()
i want the game to look worse lol.
really, i just have an idea for how the game should look and that is one which has a fixed window size. so either i implement the Funny Technique where i force the window size to be a integer multiple (which is possible and very close anyways to how UpdateDPI works) or i letterbox or i do the Funnier Technique which is this. i've decided to go with this